requirements.txt
appnope==0.1.0
astroid==2.2.5
attrs==19.2.0
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
flake8==3.7.8
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.1
Jinja2==2.10.3
json5==0.8.5
jsonschema==3.0.2
jupyter-client==5.3.3
jupyter-core==4.5.0
jupyterlab==1.1.4
jupyterlab-server==1.0.6
lazy-object-proxy==1.4.1
lxml==4.3.4
MarkupSafe==1.1.1
mccabe==0.6.1
mecab-python3==0.996.2
mistune==0.8.4
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.4.2
pylint==2.3.1
pyrsistent==0.15.4
python-dateutil==2.8.0
PyYAML==5.1.2
pyzmq==18.1.0
Send2Trash==1.5.0
six==1.12.0
soupsieve==1.9.2
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
tqdm==4.32.2
traitlets==4.3.3
typed-ast==1.4.0
wcwidth==0.1.7
webencodings==0.5.1
wrapt==1.11.2
